+++ rusoto/core/src/serialization.rs
@@ -2,11 +2,12 @@ use std::error::Error;
 use std::fmt::{Formatter, Result as FmtResult};
 
 use base64;
-use serde::de::{Error as SerdeError, Visitor};
-use serde::{Deserializer, Serializer};
+use serde::de::{Error as SerdeError, Visitor, SeqAccess};
+use serde::ser::SerializeSeq;
+use serde::{Deserialize, Deserializer, Serialize, Serializer};
 
-pub trait SerdeBlob<'de>: Sized {
-    fn deserialize_blob<D>(deserializer: D) -> Result<Self, D::Error>
+pub trait SerdeBlob: Sized {
+    fn deserialize_blob<'de, D>(deserializer: D) -> Result<Self, D::Error>
     where
         D: Deserializer<'de>;
 
@@ -32,8 +33,8 @@ impl<'de> Visitor<'de> for BlobVisitor {
     }
 }
 
-impl<'de> SerdeBlob<'de> for Vec<u8> {
-    fn deserialize_blob<D>(deserializer: D) -> Result<Vec<u8>, D::Error>
+impl SerdeBlob for Vec<u8> {
+    fn deserialize_blob<'de, D>(deserializer: D) -> Result<Vec<u8>, D::Error>
     where
         D: Deserializer<'de>,
     {
@@ -72,8 +73,8 @@ impl<'de> Visitor<'de> for OptionalBlobVisitor {
     }
 }
 
-impl<'de> SerdeBlob<'de> for Option<Vec<u8>> {
-    fn deserialize_blob<D>(deserializer: D) -> Result<Option<Vec<u8>>, D::Error>
+impl SerdeBlob for Option<Vec<u8>> {
+    fn deserialize_blob<'de, D>(deserializer: D) -> Result<Option<Vec<u8>>, D::Error>
     where
         D: Deserializer<'de>,
     {
@@ -91,12 +92,134 @@ impl<'de> SerdeBlob<'de> for Option<Vec<u8>> {
     }
 }
 
+struct DeserializeWrapper<T>(T);
+
+impl<'de, T> Deserialize<'de> for DeserializeWrapper<T>
+where
+    T: SerdeBlob,
+{
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        Ok(DeserializeWrapper(T::deserialize_blob(deserializer)?))
+    }
+}
+
+struct SerializeWrapper<'a, T>(&'a T);
+
+impl<'a, T> Serialize for SerializeWrapper<'a, T>
+where
+    T: SerdeBlob,
+{
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        T::serialize_blob(self.0, serializer)
+    }
+}
+
+pub trait SerdeBlobList: Sized {
+    fn deserialize_blob_list<'de, D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>;
+
+    fn serialize_blob_list<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer;
+}
+
+struct BlobListVisitor;
+
+impl<'de> Visitor<'de> for BlobListVisitor {
+    type Value = Vec<Vec<u8>>;
+
+    fn expecting(&self, formatter: &mut Formatter) -> FmtResult {
+        write!(formatter, "a list of blobs")
+    }
+
+    fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
+    where
+        A: SeqAccess<'de>,
+    {
+        let mut blob_list = Vec::new();
+        while let Some(DeserializeWrapper(blob)) = seq.next_element()? {
+            blob_list.push(blob);
+        }
+        Ok(blob_list)
+    }
+}
+
+impl SerdeBlobList for Vec<Vec<u8>> {
+    fn deserialize_blob_list<'de, D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        deserializer.deserialize_seq(BlobListVisitor)
+    }
+
+    fn serialize_blob_list<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        let mut seq = serializer.serialize_seq(Some(self.len()))?;
+        for blob in self.iter() {
+            seq.serialize_element(&SerializeWrapper(blob))?;
+        }
+        seq.end()
+    }
+}
+
+struct OptionalBlobListVisitor;
+
+impl<'de> Visitor<'de> for OptionalBlobListVisitor {
+    type Value = Option<Vec<Vec<u8>>>;
+
+    fn expecting(&self, formatter: &mut Formatter) -> FmtResult {
+        write!(formatter, "an optional blob list")
+    }
+
+    fn visit_none<E>(self) -> Result<Self::Value, E>
+    where
+        E: SerdeError,
+    {
+        Ok(None)
+    }
+
+    fn visit_some<D>(self, deserializer: D) -> Result<Self::Value, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        Ok(Some(SerdeBlobList::deserialize_blob_list(deserializer)?))
+    }
+}
+
+impl SerdeBlobList for Option<Vec<Vec<u8>>> {
+    fn deserialize_blob_list<'de, D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        deserializer.deserialize_option(OptionalBlobListVisitor)
+    }
+
+    fn serialize_blob_list<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        match *self {
+            Some(ref list) => SerdeBlobList::serialize_blob_list(list, serializer),
+            None => serializer.serialize_none(),
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     extern crate serde;
     extern crate serde_json;
 
-    use super::SerdeBlob;
+    use super::{SerdeBlob, SerdeBlobList};
 
     #[test]
     fn serialize_optional_blob_when_none() {
@@ -181,7 +304,45 @@ mod tests {
         assert_eq!("hello world!".as_bytes().to_vec(), deserialized);
     }
 
-    fn serialize_blob_helper<'de, B: SerdeBlob<'de>>(blob: B) -> String {
+    #[test]
+    fn serialize_blob_list_when_empty() {
+        let blob_list = vec![];
+        let serialized = serialize_blob_list_helper(blob_list);
+
+        assert_eq!("[]", serialized);
+    }
+
+    #[test]
+    fn serialize_blob_list_when_has_content() {
+        let blob_list = vec![
+            "Sunny".as_bytes().to_vec(),
+            "Rainy".as_bytes().to_vec(),
+            "Snowy".as_bytes().to_vec()
+        ];
+        let serialized = serialize_blob_list_helper(blob_list);
+
+        assert_eq!("[\"U3Vubnk=\",\"UmFpbnk=\",\"U25vd3k=\"]", serialized);
+    }
+
+    #[test]
+    fn deserialize_blob_list_when_empty() {
+        let deserialized: Vec<Vec<u8>> = deserialize_blob_list_helper("[]").unwrap();
+
+        assert_eq!(Vec::<Vec<u8>>::new(), deserialized);
+    }
+
+    #[test]
+    fn deserialize_blob_list_when_has_content() {
+        let deserialized: Vec<Vec<u8>> = deserialize_blob_list_helper("[\"U3Vubnk=\",\"UmFpbnk=\",\"U25vd3k=\"]").unwrap();
+
+        assert_eq!(vec![
+            "Sunny".as_bytes().to_vec(),
+            "Rainy".as_bytes().to_vec(),
+            "Snowy".as_bytes().to_vec()
+        ], deserialized);
+    }
+
+    fn serialize_blob_helper<B: SerdeBlob>(blob: B) -> String {
         let mut serialized_data = Vec::new();
         {
             let mut json_serializer = serde_json::Serializer::new(&mut serialized_data);
@@ -190,8 +351,8 @@ mod tests {
         String::from_utf8_lossy(&serialized_data).into_owned()
     }
 
-    fn deserialize_blob_helper<'de, B: SerdeBlob<'de>>(
-        s: &'de str,
+    fn deserialize_blob_helper<B: SerdeBlob>(
+        s: &str,
     ) -> Result<
         B,
         <&mut serde_json::de::Deserializer<serde_json::de::StrRead> as serde::Deserializer>::Error,
@@ -199,5 +360,25 @@ mod tests {
         let reader = serde_json::de::StrRead::new(s);
         let mut deserializer = serde_json::de::Deserializer::new(reader);
         B::deserialize_blob(&mut deserializer)
+    }
+
+    fn serialize_blob_list_helper<B: SerdeBlobList>(blob_list: B) -> String {
+        let mut serialized_data = Vec::new();
+        {
+            let mut json_serializer = serde_json::Serializer::new(&mut serialized_data);
+            blob_list.serialize_blob_list(&mut json_serializer).unwrap();
+        }
+        String::from_utf8_lossy(&serialized_data).into_owned()
+    }
+
+    fn deserialize_blob_list_helper<B: SerdeBlobList>(
+        s: &str,
+    ) -> Result<
+        B,
+        <&mut serde_json::de::Deserializer<serde_json::de::StrRead> as serde::Deserializer>::Error,
+    > {
+        let reader = serde_json::de::StrRead::new(s);
+        let mut deserializer = serde_json::de::Deserializer::new(reader);
+        B::deserialize_blob_list(&mut deserializer)
     }
 }